janitor: drop
instead use webhooks which are triggered via an external builder (i.e.
buildbot or CI) or run the script `update_all_targets.py` in a cron job.

Scheduling within RQ was somehow no fun and caused strangely dying jobs.

While at it, use `dotenv` to simplify things.
