+++ .gitignore
@@ -1,28 +1,29 @@
-*$py.class
-*.pub
-*.py[cod]
-*.sec
-*.sig
-*.swp
-.DS_Store
-.coverage
-.coverage.*
-.direnv/
-.venv/
-.vscode/
-__pycache__/
-asu.egg-info/
-asu/__pycache__/
-cache/
-config.py
-coverage.xml
-dist/
-htmlcov/
-instance/
-json/
-poetry.lock
-public/
-redis/
-site/
-store/
 var/
+store/
+site/
+redis/
+public/
+poetry.lock
+json/
+instance/
+htmlcov/
+dist/
+coverage.xml
+config.py
+cache/
+asu/__pycache__/
+asu.egg-info/
+*$py.class
+*.swp
+*.sig
+*.sec
+*.py[cod]
+*.pub
+.vscode/
+.venv/
+.env
+.DS_Store
+.direnv/
+.coverage.*
+.coverage
+__pycache__/
