+++ README.md
@@ -77,9 +77,9 @@ Once Podman works, install `podman-compose`:
 Now it's possible to run all services via `podman-compose`:
 
     # where to store images and json files
-    export PUBLIC_PATH=$(pwd)/public
+    echo "PUBLIC_PATH=$(pwd)/public" > .env
     # absolute path to podman socket mounted into worker containers
-    export CONTAINER_SOCK=/run/user/1001/podman/podman.sock
+    echo "CONTAINER_SOCK=/run/user/1001/podman/podman.sock" >> .env
     podman-compose up -d
 
 This will start the server, the Podman API container and two workers. The first
