+++ pyproject.toml
@@ -10,14 +10,15 @@ include = ["asu/branches.yml"]
 [tool.poetry.dependencies]
 python = "^3.11"
 PyNaCl = "^1.5.0"
-redis = ">=4.5.1,<6.0.0"
 requests = "^2.28.2"
 rq = "^1.13.0"
 connexion = {extras = ["swagger-ui"], version = "^2.14.2"}
-prometheus-client = ">=0.13.1,<0.21.0"
 gunicorn = ">=20.1,<23.0"
 podman = ">=4.4.1,<6.0.0"
 setuptools = ">=69.0.3,<71.0.0"
+python-dotenv = "^1.0.1"
+redis = "^5.0.5"
+prometheus-client = "^0.20.0"
 
 [tool.poetry.group.dev.dependencies]
 ruff = ">=0.2.2,<0.5.0"
