+++ tests/upstream/snapshots/targets/testtarget/testsubtarget/openwrt-imagebuilder-testtarget-testsubtarget.Linux-x86_64/Dockerfile
@@ -1,0 +1,12 @@
+FROM alpine
+
+RUN apk add make bash
+
+RUN adduser -D builder -h /builder/
+
+USER builder 
+
+ADD ./ /builder/
+
+WORKDIR /builder/
+
